Revert "improvement: individually buildable back, batch, preview"
This reverts commit dd9f6ab256926874df803767a14078ba5e1af3d2 due to
absolutely miserable build times for development.
