+++ Cargo.toml
@@ -1,4 +1,9 @@
 [workspace]
 resolver = "2"
 
-members = ["clovers-back", "clovers-batch", "clovers-preview"]
+members = [
+  "clovers-back",
+  "clovers-batch",
+  "clovers-preview",
+  "clovers-svc-common",
+]
